Implement the routes setter and encapsulate the reply procedure
Function SET-ROUTE is the function to be used for binding handler name
with specific HTTP method and path. It's just a function form of the
code above it. Function REPLY is also a thin abstraction and wrapper
of function application of WRITE-HEADER and PRINC for returnning
response message to client.